Fixes #3. Added a method that does tree initialization and calls the search algorithm. Extracted method for checking pathfinder result validity.
Concluded that the tree itself solves the problem of finding all paths between any two words. The way the tree works makes any attempt at navigating the tree and searching for paths between any two words (if the first word is not in the root) inefficient. It's better to just generate a new tree and search through it.